la RicercaServlet restituisce tutte le informazioni inerenti alla pagina con tutti i fornitori e tutti i prezzi dei fornitori per ogni prodotto.

+++ src/main/java/beans/FornitoreConProdotto.java
@@ -2,15 +2,15 @@ package beans;
 
 public class FornitoreConProdotto {
 
-    private int prezzoVendita;
+    private float prezzoVendita;
 
     private Fornitore fornitore;
 
-    public int getPrezzoVendita() {
+    public float getPrezzoVendita() {
         return prezzoVendita;
     }
 
-    public void setPrezzoVendita(int prezzoVendita) {
+    public void setPrezzoVendita(float prezzoVendita) {
         this.prezzoVendita = prezzoVendita;
     }
